Add in functionality for export all, export contact messages, and export conversation buttons. Includes good error checking at all points and enabling/ disabling of export buttons when they should/should not work.
Default format is the same as smstotext files. Exporting a contact's messages
omits the contact's name and number and names the file after the contact by
default. Exporting a conversation omits the date on each line and adds the long
date at the beginning of the file. It would be nice to have the date in the file
name too, but I'm not sure how to do that in a good way.