Added the actual sync step to the app. Texts from the network link are now loaded into the main database, with contacts list computed properly. Still missing the final step of the sync and some error checking.
Note that the sync across thread boundaries is accomplished with a shared
private variable in the MessageDatabase class and 2 functions.
ReadFromNetworkInput is meant to be called from the network thread only, and
TryUpdateFromNetwork is meant to be called from the UI thread only. The main
window has a timer that fires every 2 seconds to check for new messages in the
shared variable - I couldn't find a way to trigger code on the UI thread from
the network thread.