Added the final step to the sync along with a few extra checks and safeguards. Now checks the number of messages properly decoded against the number of messages sent, and doesn't add any of them if they don't match.
Also got rid of the irrelevant import from droid button and code. Don't know
why that was still in there.

Still needs some cleanup and removal of debugging stuff after network operation
is confirmed to be working right, and need to let it save the file again. Also
need to figure out a good way to enter, store, and retrieve the sync password
and port specification.